requirements.txt
pydantic-ai==0.0.22
asyncpg==0.30.0
pgvector==0.3.6
python-dotenv==1.0.1
openai==1.58.1
numpy==2.2.0
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
psycopg2-binary==2.9.10
sentence-transformers==3.3.1
